requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
certifi==2024.8.30
commentjson==0.9.0
charset-normalizer==3.4.0
distro==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.27.2
idna==3.10
jiter==0.7.1
openai==1.54.4
outcome==1.3.0.post0
pydantic==2.9.2
pydantic_core==2.23.4
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.4
selenium==4.26.1
sniffio==1.3.1
sortedcontainers==2.4.0
tqdm==4.67.0
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
undetected-chromedriver==3.5.5
urllib3==2.5.0
websocket-client==1.8.0
websockets==14.1
wsproto==1.2.0
pillow==11.3.0

